+++ config/decoder.go
@@ -73,6 +73,13 @@ func convertValueFromStruct(value interface{}, targetType reflect.Type, fieldTyp
 	return nil
 }
 
+func addSeparator(key string) string {
+	if key != "" {
+		key += _separator
+	}
+	return key
+}
+
 type decoder struct {
 	*Value
 	m map[interface{}]struct{}
@@ -125,7 +132,7 @@ func (d *decoder) sequence(childKey string, value reflect.Value) error {
 
 	// start looking for child values.
 	elementType := derefType(valueType).Elem()
-	childKey += _separator
+	childKey = addSeparator(childKey)
 
 	for ai := 0; ; ai++ {
 		arrayKey := childKey + strconv.Itoa(ai)
@@ -164,7 +171,7 @@ func (d *decoder) array(childKey string, value reflect.Value) error {
 
 	// start looking for child values.
 	elementType := derefType(valueType).Elem()
-	childKey += _separator
+	childKey = addSeparator(childKey)
 
 	for ai := 0; ai < value.Len(); ai++ {
 		arrayKey := childKey + strconv.Itoa(ai)
@@ -205,13 +212,18 @@ func (d *decoder) mapping(childKey string, value reflect.Value, def string) erro
 	// child yamlNode parsed from yaml file is of type map[interface{}]interface{}
 	// type casting here makes sure that we are iterating over a parsed map.
 	if v, ok := val.(map[interface{}]interface{}); ok {
-		childKey += _separator
+		childKey = addSeparator(childKey)
+
 		for key := range v {
-			mapKey := childKey + fmt.Sprintf("%v", key)
+			subKey := fmt.Sprintf("%v", key)
+			if subKey == "" {
+				return fmt.Errorf("empty key leads to ambiguity for path: %q", childKey)
+			}
+
 			itemValue := reflect.New(valueType.Elem()).Elem()
 
 			// Try to unmarshal value and save it in the map.
-			if err := d.unmarshal(mapKey, itemValue, def); err != nil {
+			if err := d.unmarshal(childKey+subKey, itemValue, def); err != nil {
 				return err
 			}